fragment: fix NPE if MainActivity be recreated
To use FragmentStatePagerAdapter in correct way. If MainActivity is
destroyed by system when we go back from other activity, we will get
NPE.

The problem is we maintained another fragment in Adapter, but
FragmentStatePagerAdapter asks FragmentManager to restore/re-create
another Fragment, hence the fragment we hold in Adapter is not the real
one which be used by system.

refer: https://stackoverflow.com/a/15261142/236743